improve packaging
