+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
         </activity>
         <activity android:name=".OtherActivity"
             android:theme="@style/Theme.AppCompat.NoActionBar"/>
-        <activity android:name=".JetpackComposeActivity"
+        <activity android:name=".compose.JetpackComposeActivity"
             android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
